requirements.txt
aiogram
google-genai
python-dotenv
flask